requirements.txt
# 对冲策略交易机器人依赖包
# Python版本要求: >=3.7

# 币安API客户端
python-binance>=1.0.0

# HTTP请求库（用于钉钉通知）
requests>=2.25.0

# 数据处理库（用于EMA等计算）
pandas>=1.3.0

